Fixed bug where saveItem would fail because it tried to add the item to the keychain when it already existed.
itemForKey: returns nil if it cannot unarchive the NSData that is
stored in the keychain. This caused an issue because saveItem: used the
returned value of itemForKey: to determine if an item existed in the
keychain.

itemForKey: could return nil when an item actually did exist in the
keychain and because saveItem: thought no item existed it would attempt
to add one to the keychain would result in an error.

saveItem: now properly checks if the item exists by checking the raw
item attributes and data that is returned when querying the keychain.