Added NSError parameters to all method calls.
The introduction of background fetching in iOS 7 means a lot more apps
will be executing code when the device is locked. This can cause
problems because by default items in the keychain are only accessible
when the device is unlocked.

All methods will now return back NSError objects which will wrap any
errors returned from Keychain Services calls. Error codes will match
those found in SecBase.h.

A check for the error code errSecInteractionNotAllowed can be used to
determine if the item returned from the keychain is nil because the
device was locked.