itemForKey: now returns nil and an error if the item cannot be deserialized successfully.
In theory if a user changed the structure of a class that had been
encoded to the keychain the next time that item was loaded an exception
could be thrown.

Rather then rely on this exception to let the user know that something
has to be done FDKeychain was modified to catch this exception and
return an error object instead.

This gives users the ability to recover from this crash instead of
forcing them to modify their classes until the item can be successfully
deserialized again.